+++ package.json
@@ -26,7 +26,6 @@
     "form-data": "^4.0.0",
     "fs": "^0.0.1-security",
     "hummus-recipe": "^2.0.1",
-    "lodash": "^4.17.21",
     "markdown-it": "^13.0.1",
     "markdown-it-abbr": "^1.0.4",
     "markdown-it-container": "^3.0.0",
@@ -53,14 +52,14 @@
   "devDependencies": {
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
-    "@types/hummus-recipe": "^1.8.5",
+    "@types/hummus-recipe": "^1.8.8",
     "@types/jest": "^29.5.0",
-    "@types/lodash": "^4.14.195",
     "@types/markdown-it": "^12.2.3",
     "@types/markdown-it-container": "^2.0.5",
     "@types/markdown-it-emoji": "^2.0.2",
     "@types/markdown-it-footnote": "^3.0.0",
     "@types/mime-types": "^2.1.1",
+    "@types/node": "^20.12.7",
     "@types/node-fetch": "^2.6.4",
     "@types/nodemailer": "^6.4.7",
     "@types/sqlite3": "^3.1.8",
