+++ src/core/helpers/index.ts
@@ -14,5 +14,4 @@ export * from './fillArrayFromTo'
 export * from './toFixed'
 export * from './calculator'
 export * from './getScoresSummary'
-export * from './LoggerService'
 export * from './isNotEmpty'
