+++ src/core/helpers/patchConditionalInScoreReport.ts
@@ -1,11 +1,10 @@
-import { cloneDeep } from 'lodash'
 import { IActivityScoresAndReportsConditionalLogic, IActivityScoresAndReportsScores } from '../interfaces'
 
 export function patchConditionalInScoreReport(
   conditional: IActivityScoresAndReportsConditionalLogic,
   report: IActivityScoresAndReportsScores,
 ): IActivityScoresAndReportsConditionalLogic {
-  const condClone = cloneDeep(conditional)
+  const condClone = structuredClone(conditional)
 
   for (const condition of condClone.conditions) {
     if (report.id.endsWith(condition.itemName)) {
