+++ src/modules/report/report.controller.ts
@@ -8,7 +8,8 @@ import { v4 } from 'uuid'
 import { decryptData } from '../../encryption'
 import { ActivityResponse, SendPdfReportResponse } from '../../core/interfaces'
 import { getAppletKeys } from '../../db'
-import { convertMarkdownToHtml, logger } from '../../core/helpers'
+import { convertMarkdownToHtml } from '../../core/helpers'
+import { logger } from '../../core/services/LoggerService'
 import { AppletEntity } from '../../models'
 import { getCurrentCount, convertHtmlToPdf, watermarkPDF, encryptPDF, getPDFPassword } from '../../pdf-utils'
 import { SendPdfReportRequest, SendPdfReportRequestPayload } from './types'
